requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.0
certifi==2022.6.15
charset-normalizer==2.0.12
greenlet==1.1.2
Mako==1.2.0
packaging==21.3
pluggy==1.0.0
pyTelegramBotAPI==4.6.0
pytest==7.1.2
pytest-asyncio==0.18.3
python-dateutil==2.8.2
python-decouple==3.6
python-telegram-bot-calendar==1.0.5
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
typing_extensions==4.2.0
